Projectie shoots the correct way the player is facing

+++ Assets/General Scripts/buttonController.cs
@@ -59,7 +59,7 @@ public class buttonController : MonoBehaviour
 
     void Update()
     {
-        GameManager.Instance.mult = (1 + GameManager.Instance.streak * 0.1);
+        GameManager.Instance.mult = (float)(1 + GameManager.Instance.streak * 0.1);
 
         timeSinceLastNoteHit += Time.deltaTime;
